data pipeline reformat
